+++ allauth/socialaccount/providers/oauth2/client.py
@@ -11,7 +11,7 @@ class OAuth2Error(Exception):
 
 
 class OAuth2Client(object):
-    CLIENT_ID_PARAMETER = "client_id"
+    client_id_parameter = "client_id"
 
     def __init__(
         self,
@@ -39,7 +39,7 @@ class OAuth2Client(object):
     def get_redirect_url(self, authorization_url, scope, extra_params):
         scope = self.scope_delimiter.join(set(scope))
         params = {
-            self.CLIENT_ID_PARAMETER: self.consumer_key,
+            self.client_id_parameter: self.consumer_key,
             "redirect_uri": self.callback_url,
             "scope": scope,
             "response_type": "code",
@@ -61,7 +61,7 @@ class OAuth2Client(object):
             auth = None
             data.update(
                 {
-                    self.CLIENT_ID_PARAMETER: self.consumer_key,
+                    self.client_id_parameter: self.consumer_key,
                     "client_secret": self.consumer_secret,
                 }
             )