+++ allauth/socialaccount/providers/tiktok/client.py
@@ -1,5 +1,5 @@
 from allauth.socialaccount.providers.oauth2.client import OAuth2Client
 
 
-class TiktokOAuth2Client(OAuth2Client):
-    CLIENT_ID_PARAMETER = "client_key"
+class TikTokOAuth2Client(OAuth2Client):
+    client_id_parameter = "client_key"