+++ allauth/socialaccount/providers/tiktok/provider.py
@@ -1,10 +1,10 @@
 from allauth.socialaccount.providers.base import ProviderAccount
 from allauth.socialaccount.providers.oauth2.provider import OAuth2Provider
-from allauth.socialaccount.providers.tiktok.scopes import TiktokScope
-from allauth.socialaccount.providers.tiktok.views import TiktokOAuth2Adapter
+from allauth.socialaccount.providers.tiktok.scope import TikTokScope
+from allauth.socialaccount.providers.tiktok.views import TikTokOAuth2Adapter
 
 
-class TiktokAccount(ProviderAccount):
+class TikTokAccount(ProviderAccount):
 
     def get_username(self):
         return self.account.extra_data.get("username")
@@ -20,29 +20,29 @@ class TiktokAccount(ProviderAccount):
 
     def to_str(self):
         username = self.get_username()
-        return username or super(TiktokAccount, self).to_str()
+        return username or super().to_str()
 
 
-class TiktokProvider(OAuth2Provider):
+class TikTokProvider(OAuth2Provider):
     id = "tiktok"
-    name = "Tiktok"
-    account_class = TiktokAccount
-    oauth2_adapter_class = TiktokOAuth2Adapter
+    name = "TikTok"
+    account_class = TikTokAccount
+    oauth2_adapter_class = TikTokOAuth2Adapter
     pkce_enabled_default = False
 
     def extract_uid(self, data):
         return str(data["open_id"])
 
     def extract_common_fields(self, data):
+        # TikTok does not provide an email address
         return {
             "username": data.get("username") or data.get("display_name"),
             "name": data.get("display_name"),
-            "email": None,  # Tiktok does not provide an email
         }
 
     def get_default_scope(self):
-        # Requires LogitKit and Scopes with user.info.basic and user.info.profile enabled
-        return [TiktokScope.user_basic_info, TiktokScope.user_profile_info]
+        # Requires LoginKit and Scopes with user.info.basic and user.info.profile enabled
+        return [TikTokScope.user_info_basic, TikTokScope.user_info_profile]
 
 
-provider_classes = [TiktokProvider]
+provider_classes = [TikTokProvider]