+++ allauth/socialaccount/providers/tiktok/scope.py
@@ -1,0 +1,6 @@
+from enum import StrEnum
+
+
+class TikTokScope(StrEnum):
+    user_info_basic = "user.info.basic"
+    user_info_profile = "user.info.profile"