+++ allauth/socialaccount/providers/tiktok/scopes.py
@@ -1,6 +1,0 @@
-from enum import StrEnum
-
-
-class TiktokScope(StrEnum):
-    user_basic_info = "user.info.basic"
-    user_profile_info = "user.info.profile"