+++ allauth/socialaccount/providers/tiktok/urls.py
@@ -1,6 +1,6 @@
 from allauth.socialaccount.providers.oauth2.urls import default_urlpatterns
 
-from .provider import TiktokProvider
+from .provider import TikTokProvider
 
 
-urlpatterns = default_urlpatterns(TiktokProvider)
+urlpatterns = default_urlpatterns(TikTokProvider)